requirements.txt
Flask==2.3.2
opencv-python==4.8.0.76
face-recognition==1.3.0
numpy==1.23.5
Pillow==9.5.0
mediapipe==0.10.0